requirements.txt
torch 
torchvision 
diffusers 
fastapi 
uvicorn 
pillow 
opencv-contrib-python
accelerate
xformers
mediapipe
controlnet_aux
transformers
streamlit